requirements.txt
# PupRing AI Python Services Requirements

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2

# Background Removal
rembg==2.0.50

# Image Enhancement
scikit-image==0.22.0
imageio==2.33.1

# HTTP Requests
requests==2.31.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0

# Development
pytest==7.4.3